Leaderboards

+++ database/migrations/0001_01_01_000000_create_users_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->timestamp('email_verified_at')->nullable();
             $table->string('name')->unique();
             $table->string('password');
-            $table->decimal('balance', 10, 2)->default(0.00)->unsigned();
+            $table->integer('balance')->default(0.00)->unsigned();
             $table->rememberToken();
             $table->timestamps();
         });
